Implement a basic installer using Wix v3
This started off with the template from the VS 2019 extension, then
taking a lot of inspiration from `kurtanr`'s `WiXInstallerExamples`
project to put it all together.

We can now generate both x86 and x64 installers that put the file in the
correct path and set the environment variable correctly for each.

The bundled binaries come directly from 1Password's release page.

https://marketplace.visualstudio.com/items?itemName=WixToolset.WixToolsetVisualStudio2019Extension
https://github.com/kurtanr/WiXInstallerExamples/tree/main/02_x86_x64_Installer
https://app-updates.agilebits.com/product_history/CLI2